Extract reusable functions
